Rename sbom module to report
