+++ dfetch/__main__.py
@@ -14,7 +14,7 @@ import dfetch.commands.freeze
 import dfetch.commands.import_
 import dfetch.commands.init
 import dfetch.commands.list
-import dfetch.commands.sbom
+import dfetch.commands.report
 import dfetch.commands.update
 import dfetch.commands.validate
 import dfetch.log
@@ -45,7 +45,7 @@ def create_parser() -> argparse.ArgumentParser:
     dfetch.commands.import_.Import.create_menu(subparsers)
     dfetch.commands.init.Init.create_menu(subparsers)
     dfetch.commands.list.List.create_menu(subparsers)
-    dfetch.commands.sbom.Sbom.create_menu(subparsers)
+    dfetch.commands.report.Report.create_menu(subparsers)
     dfetch.commands.update.Update.create_menu(subparsers)
     dfetch.commands.validate.Validate.create_menu(subparsers)
